Port src/marks/{box,bollinger,linearRegression} from JS to TS (stage 1)
Inline hand-written .d.ts types into the source files, deleting the
.js/.d.ts pairs. No behaviour changes; runtime preserved verbatim with
parameter annotations and casts where the existing shared .d.ts files
omit declarations for runtime-only helpers.
